+++ src/union_physics/v5.rs
@@ -1,0 +1,2 @@
+// These formats are identical
+pub use super::v3::CSGPHS3 as CSGPHS5;
