+++ src/union_physics/v6.rs
@@ -1,0 +1,9 @@
+use super::v3::{Mesh, PhysicsInfo};
+
+#[binrw::binrw]
+#[brw(little)]
+#[derive(Debug, Clone)]
+pub struct CSGPHS6 {
+	pub physics_info: PhysicsInfo,
+	pub mesh: Mesh,
+}
